+++ CHANGELOG.rst
@@ -24,7 +24,8 @@ Deprecations:
 Changes:
 ^^^^^^^^
 
-*none*
+- ``structlog.dev.ConsoleRenderer`` now has ``sort_keys`` boolean parameter which controls whether to sort keys when formatting keys output. `True` by default.
+  `#331 <https://github.com/hynek/structlog/pull/358>`_
 
 
 ----