+++ CHANGELOG.rst
@@ -26,7 +26,8 @@ Changes:
 
 - ``structlog.dev.ConsoleRenderer`` now has ``sort_keys`` boolean parameter that allows to disable the sorting of keys on output.
   `#358 <https://github.com/hynek/structlog/pull/358>`_
-
+- ``structlog.processors.TimeStamper`` now works well with FreezeGun even when it gets applied before the loggers are configured.
+  `#364 <https://github.com/hynek/structlog/pull/364>`_
 
 ----