+++ CHANGELOG.rst
@@ -5,7 +5,7 @@ Versions are year-based with a strict backward compatibility policy.
 The third digit is only for regressions.
 
 
-21.3.0 (UNRELEASED)
+21.3.0 (2021-11-20)
 -------------------